Image Colour Palette Generator/requirements.txt
Flask==3.0.3
numpy==2.1.2
Pillow==11.0.0
scikit_learn==1.5.2
